backend/requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
opencv-python
tensorflow
tensorflow-io
pillow
pickle-mixin
